Move: 랜덤뽑기게임 package day03 -> day03.mission0
