+++ week01/src/day03/mission0/RandomGacha.java
@@ -1,4 +1,4 @@
-package day03;
+package day03.mission0;
 
 import java.util.*;
